% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.data.frame.dfa}
\alias{as.data.frame.dfa}
\title{Таблица влияния факторов}
\usage{
\method{as.data.frame}{dfa}(x, ...)
}
\arguments{
\item{x}{Объект класса \code{dfa}.}

\item{...}{Не используется.}
}
\value{
\code{data.frame} с колонками \code{factor}, \code{effect}, \code{share}, \code{formula}.
}
\description{
Таблица влияния факторов
}
