% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.dfa}
\alias{plot.dfa}
\title{Каскадная диаграмма (водопад) влияния факторов}
\usage{
\method{plot}{dfa}(
  x,
  col_up = "#2E7D32",
  col_down = "#C62828",
  col_total = "#546E7A",
  main = NULL,
  digits = 2,
  ...
)
}
\arguments{
\item{x}{Объект класса \code{dfa}.}

\item{col_up, col_down, col_total}{Цвета роста, снижения и итоговых столбцов.}

\item{main}{Заголовок.}

\item{digits}{Знаков после запятой в подписях.}

\item{...}{Передаётся в \code{\link[graphics:plot.window]{graphics::plot.window()}}.}
}
\description{
Каскадная диаграмма (водопад) влияния факторов
}
\examples{
res <- chain_substitution(V ~ N * W, c(N = 100, W = 10), c(N = 120, W = 12))
plot(res)
}
