% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dfa}
\alias{print.dfa}
\title{Печать результата факторного анализа}
\usage{
\method{print}{dfa}(x, digits = 2, details = TRUE, ...)
}
\arguments{
\item{x}{Объект класса \code{dfa}.}

\item{digits}{Знаков после запятой.}

\item{details}{Показывать ли расчётные таблицы метода.}

\item{...}{Не используется.}
}
\description{
Печать результата факторного анализа
}
